+++ selfdrive/ui/lib/api_helpers.py
@@ -9,6 +9,8 @@ from openpilot.common.swaglog import cloudlog
 from openpilot.common.params import Params
 from openpilot.selfdrive.ui.ui_state import ui_state, device
 
+from system.athena.registration import UNREGISTERED_DONGLE_ID
+
 TOKEN_EXPIRY_HOURS = 2
 
 
@@ -38,49 +40,55 @@ class RequestRepeater:
     self._request_done_callbacks: list[Callable[[str, bool], None]] = []
     self.add_request_done_callback(self._handle_reply)
 
-    self._prev_request_text = None
+    self._prev_response_text = None
     self._lock = threading.Lock()
     self._running = False
     self._thread = None
     self._data = None
-    self._last_request_time = 0.0
     self._params = Params()
-
-    self.start()
 
   def add_request_done_callback(self, callback: Callable[[str, bool], None]):
     self._request_done_callbacks.append(callback)
 
   def _handle_reply(self, response: str, success: bool):
     # Cache successful responses to params
-    if success and response != self._prev_request_text:
+    if success and response != self._prev_response_text:
       self._params.put(self._cache_key, response)
-      self._prev_request_text = response
+      self._prev_response_text = response
 
   def start(self):
+    if self._thread and self._thread.is_alive():
+      return
     self._running = True
-    self._thread = threading.Thread(target=self._run_thread, daemon=True)
+    self._thread = threading.Thread(target=self._worker_thread, daemon=True)
     self._thread.start()
 
-  def _run_thread(self):
+  def stop(self):
+    self._running = False
+    if self._thread and self._thread.is_alive():
+      self._thread.join(timeout=1.0)
+
+  def _worker_thread(self):
     while self._running:
-      now = time.monotonic()
-      if now - self._last_request_time >= self._period:
-        self._last_request_time = now
+      active_request = False  # TODO: this
+      if not ui_state.started and device.is_awake() and not active_request:
+        print('RUNNING REQUEST')
+        self._send_request()
+      else:
+        print('SKIPPING REQUEST')
 
-        active_request = False   # TODO: this
-        if not ui_state.started and device.is_awake() and not active_request:
-          print('RUNNING REQUEST')
-          self._send_request()
-        else:
-          print('SKIPPING REQUEST')
-
-      time.sleep(self.SLEEP_INTERVAL)
+      for _ in range(int(self.FETCH_INTERVAL / self.SLEEP_INTERVAL)):
+        if not self._running:
+          break
+        time.sleep(self.SLEEP_INTERVAL)
 
   def _send_request(self):
+    if not self._dongle_id or self._dongle_id == UNREGISTERED_DONGLE_ID:
+      return
+
     try:
       identity_token = get_token(self._dongle_id)
-      response = api_get(f"v1.1/devices/{self._dongle_id}", timeout=self.API_TIMEOUT, access_token=identity_token)
+      response = api_get(self._request_route, timeout=self.API_TIMEOUT, access_token=identity_token)
 
       if response.status_code == 200:
         for callback in self._request_done_callbacks:
@@ -90,4 +98,7 @@ class RequestRepeater:
           callback(response.text, False)
 
     except Exception as e:
-      cloudlog.error(f"Failed to fetch prime status: {e}")
+      cloudlog.error(f"Failed to send request to {self._request_route}: {e}")
+
+  def __del__(self):
+    self.stop()
