+++ selfdrive/ui/lib/prime_state.py
@@ -63,22 +63,6 @@ class PrimeState:
     except Exception as e:
       cloudlog.error(f"Failed to parse prime status from response: {e}")
 
-  def _fetch_prime_status(self) -> None:
-    dongle_id = self._params.get("DongleId")
-    if not dongle_id or dongle_id == UNREGISTERED_DONGLE_ID:
-      return
-
-    try:
-      identity_token = get_token(dongle_id)
-      response = api_get(f"v1.1/devices/{dongle_id}", timeout=self.API_TIMEOUT, access_token=identity_token)
-      if response.status_code == 200:
-        data = response.json()
-        is_paired = data.get("is_paired", False)
-        prime_type = data.get("prime_type", 0)
-        self.set_type(PrimeType(prime_type) if is_paired else PrimeType.UNPAIRED)
-    except Exception as e:
-      cloudlog.error(f"Failed to fetch prime status: {e}")
-
   def set_type(self, prime_type: PrimeType) -> None:
     with self._lock:
       if prime_type != self.prime_type:
@@ -88,7 +72,7 @@ class PrimeState:
 
   def _worker_thread(self) -> None:
     while self._running:
-      self._fetch_prime_status()
+      # self._fetch_prime_status()
 
       for _ in range(int(self.FETCH_INTERVAL / self.SLEEP_INTERVAL)):
         if not self._running:
@@ -96,17 +80,7 @@ class PrimeState:
         time.sleep(self.SLEEP_INTERVAL)
 
   def start(self) -> None:
-    return
-    if self._thread and self._thread.is_alive():
-      return
-    self._running = True
-    self._thread = threading.Thread(target=self._worker_thread, daemon=True)
-    self._thread.start()
-
-  def stop(self) -> None:
-    self._running = False
-    if self._thread and self._thread.is_alive():
-      self._thread.join(timeout=1.0)
+    self._request_repeater.start()
 
   def get_type(self) -> PrimeType:
     with self._lock:
@@ -119,6 +93,3 @@ class PrimeState:
   def is_paired(self) -> bool:
     with self._lock:
       return self.prime_type > PrimeType.UNPAIRED
-
-  def __del__(self):
-    self.stop()
